채팅 삭제 전에 상태 확인하기

+++ commands/song_search.js
@@ -25,8 +25,6 @@ module.exports = {
         }
 
         const search = args.join(' ');
-        const resultsEmbed = new MessageEmbed().setTitle('**재생할 노래의 번호를 알려주세요.**').setDescription(`${search}의 검색 결과`).setColor('#FF9899');
-
         const filter = (await ytsr.getFilters(search)).get('Type').get('Video').url;
         const results = filter && (await ytsr(filter, { limit: 10 })).items.filter((v) => v.type == 'video'); // 영상만 가져오기
         // const results = await youtube.searchVideos(search, 10);
@@ -34,6 +32,7 @@ module.exports = {
             return message.reply('검색 내용에 해당하는 영상을 찾지 못했습니다.');
         }
 
+        const resultsEmbed = new MessageEmbed().setTitle('**재생할 노래의 번호를 알려주세요.**').setDescription(`${search}의 검색 결과`).setColor('#FF9899');
         // results.forEach((video, index) => resultsEmbed.addField(video.shortURL, `${index + 1}. ${video.title.decodeHTML().decodeHTML()}`));
         results.forEach((video, index) => resultsEmbed.addField(`https://youtu.be/${video.id}`, `${index + 1}. ${video.title}`));
 
@@ -42,24 +41,25 @@ module.exports = {
         message.channel.activeCollector = true;
         try {
             let songChoice;
-            const response = await message.channel.awaitMessages((msg) => msg.author.id == message.author.id && (songChoice = msg.content.split(',')).every((v) => !isNaN(v) && 1 <= +v && +v <= results.length), { max: 1, time: 20000, errors: ['time'] });
+            const rslt = await message.channel.awaitMessages((msg) => msg.author.id == message.author.id && (songChoice = msg.content.split(',')).every((v) => !isNaN(v) && 1 <= +v && +v <= results.length), { max: 1, time: 20000, errors: ['time'] });
 
             for (let song of songChoice) {
                 await client.commands.find((cmd) => cmd.command.includes('play')).execute(message, [resultsEmbed.fields[+song - 1].name]);
             }
 
-            delete message.channel.activeCollector;
-            resultsMessage.delete();
-            response.first().delete();
+            rslt.first().delete();
         } catch (e) {
-            if (!(e instanceof Collection)) {
+            if (e.message == 'Missing Permissions') {
+                message.channel.send('**권한이 없습니다 - [MANAGE_MESSAGES]**');
+            } else if (!(e instanceof Collection)) {
                 // awaitMessages에서 발생한 시간초과 에러는 Collection<Snowflake, Message>
                 replyAdmin(`작성자: ${message.author.username}\n방 ID: ${message.channel.id}\n채팅 내용: ${message.content}\n에러 내용: ${e}\n${e.stack ?? e._p}`);
-            } else if (e.message == 'Missing Permissions') {
-                message.channel.send('**권한이 없습니다 - [MANAGE_MESSAGES]**');
             }
+        } finally {
             delete message.channel.activeCollector;
-            resultsMessage.delete();
+            if (!resultsMessage.deleted) {
+                resultsMessage.delete();
+            }
         }
     }
 };
